Orion68 OS compiled but not tested

+++ src/env_68k/src/string/strncmp.c
@@ -2,7 +2,7 @@
 
 #include <string.h>
 
-int strncmp(const char *str1, const char *str2, int max)
+int strncmp(const char *str1, const char *str2, size_t max)
 {
 	while (max) {
 		if (*str1 > *str2)
